# 🚀 Story assets preloading improvements.