--- a/css/amp.css
+++ b/css/amp.css
@@ -862,8 +862,13 @@ amp-story .i-amphtml-loader {
   display: none !important;
 }
 
-amp-story-page:not(:first-of-type) {
-  transform: translateY(1000vh);
+/**
+ * Uses a selector that stops targeting amp-story-page elements once they have a
+ * [distance] or [active] attribute, so amp-story.css doesn't have to rely on
+ * CSS specificity hacks to override these styles after the JavaScript ran.
+ */
+amp-story-page:not(:first-of-type):not([distance]):not([active]) {
+  transform: translateY(1000vh) !important;
 }
 
 /**