--- a/extensions/amp-story/1.0/amp-story.css
+++ b/extensions/amp-story/1.0/amp-story.css
@@ -210,8 +210,12 @@ amp-story-page[active] {
  * not be automatically laid out. Max distance is set to 2 (next 2 pages) since
  * we don't want to schedule any further pages. */
 
+amp-story-page:not(:first-of-type):not([distance]):not([active]) {
+  transform: translateY(1000vh) !important;
+}
+
 amp-story-page:not([active]) {
-  transform: translateY(1000%) !important;
+  transform: translateY(1000vh) !important;
 }
 
 amp-story-page[active],