Stories assets preloading is broken in a way that if you have a 10 pages story with an image on each page, these 10 images will get loaded right away.

This PR positions the `amp-story-page` elements far away from the viewport, until the story runtime starts adding `[active]` or `[distance]` attribute and controls when elements should be built and loaded by the AMP runtime.

If the AMP runtime is loaded before the AMP Story runtime, it would start building and loading `amp-img` and `amp-video` elements. To avoid these race conditions, the CSS fix is added to both `amp.css` and `amp-story.css`.

The behavior is the following: it hides all `amp-story-pages` except for the first page, until an `[active]` or `[distance]` attribute is added once the `amp-story.buildCallback/layoutCallback` are executed.
It will start loading the first page even before it is built, and then the N+1 and N+2 pages when the `[distance]` attributes are set.

Performance measured is the time between the `amp-story.js` script onload event, and the "first page assets are loaded" event:

| Connection | Avg first page loaded time before | Avg first page loaded time after | % improvement |
| :-------------: | :-------------: | :-------------: | :-------------: |
| 3G  | 5067ms | 2804ms | -44.66% |
| 4G  | 1875ms | 898ms | -52.11% |

It's also saving a lot of bandwidth when prerendering stories from a viewer, as it only loads what's necessary.

[Demo available here](https://stamp-gmajoulet.firebaseapp.com/examples/s20/body-painting/index.html)