@gmajoulet 2019-03-16
The amp-story.css has the !important. I’m a bit scared to play with
specificity between two CSS files, but I can try to make sure it works

On Fri, Mar 15, 2019 at 8:21 PM Jon Newmuis <notifications@github.com>
wrote:

> *@newmuis* commented on this pull request.
> ------------------------------
>
> In css/amp.css
> <https://github.com/ampproject/amphtml/pull/21442#discussion_r266180037>:
>
> > @@ -862,8 +862,8 @@ amp-story .i-amphtml-loader {
>    display: none !important;
>  }
>
> -amp-story-page:not(:first-of-type):not([distance]):not([active]) {
> -  transform: translateY(1000vh) !important;
> +amp-story-page:not(:first-of-type) {
> +  transform: translateY(1000vh);
>
> Maybe we should keep the !important, since publishers could have
> transforms in their styles
>
> —
> You are receiving this because you authored the thread.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/ampproject/amphtml/pull/21442#pullrequestreview-215290348>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABbETJrw6e9XYt5hXS_u5TpPWKfWe1t6ks5vXDkigaJpZM4b3Ri8>
> .
>
-- 
Gabriel Majoulet  | UX Engineer | gmajoulet@google.com