@newmuis 2019-03-16
I think it's probably better to err on the side of leaving the `!important`, even if it means leaving the `:not([active])` and `:not([distance])`, then.  Publisher CSS sometimes has unintended consequences from overly-general selectors, and we wouldn't want that to actually have an impact on load times.